Warning Fixes

+++ Project/Content/Sample3DSceneRenderer.cpp
@@ -429,16 +429,16 @@ void Sample3DSceneRenderer::CreateDeviceDependentResources(void)
 			vertexBufferData.pSysMem = verts.data();
 			vertexBufferData.SysMemPitch = 0;
 			vertexBufferData.SysMemSlicePitch = 0;
-			CD3D11_BUFFER_DESC vertexBufferDesc(sizeof(VertexPositionUVNormal) * verts.size(), D3D11_BIND_VERTEX_BUFFER);
+			CD3D11_BUFFER_DESC vertexBufferDesc(sizeof(VertexPositionUVNormal) * (unsigned int)verts.size(), D3D11_BIND_VERTEX_BUFFER);
 			DX::ThrowIfFailed(m_deviceResources->GetD3DDevice()->CreateBuffer(&vertexBufferDesc, &vertexBufferData, &m_scene.models[0].m_vertexBuffer));
 
-			m_scene.models[0].m_indexCount = inds.size();
+			m_scene.models[0].m_indexCount = (unsigned int)inds.size();
 
 			D3D11_SUBRESOURCE_DATA indexBufferData = { 0 };
 			indexBufferData.pSysMem = inds.data();
 			indexBufferData.SysMemPitch = 0;
 			indexBufferData.SysMemSlicePitch = 0;
-			CD3D11_BUFFER_DESC indexBufferDesc(sizeof(unsigned int) * inds.size(), D3D11_BIND_INDEX_BUFFER);
+			CD3D11_BUFFER_DESC indexBufferDesc(sizeof(unsigned int) * (unsigned int)inds.size(), D3D11_BIND_INDEX_BUFFER);
 			DX::ThrowIfFailed(m_deviceResources->GetD3DDevice()->CreateBuffer(&indexBufferDesc, &indexBufferData, &m_scene.models[0].m_indexBuffer));
 		}
 	});
